update:style_logic
