+++ src/components/NewsList/styles.css
@@ -1,0 +1,24 @@
+.newsTitle {
+  margin-bottom: 10 px;
+  font-size: 20px;
+}
+
+.newsTitle:hover {
+  background-color: gray;
+}
+
+.newsPointGood {
+  color: blue;
+}
+
+.newsPointSoso {
+  color: orange;
+}
+
+.newsPointBad {
+  color: red;
+}
+
+.userName {
+  font-size: 16px;
+}
